🚚 Rename components with PascalCase convention
