+++ src/views/pages/projectDetailsPage/components/timeline/components/index.ts
@@ -1,5 +1,5 @@
-export * from './ContentArea';
-export * from './Date';
-export * from './ItemTitle';
-export * from './stateIcons';
-export * from './TimelineItem';
+export * from './ContentArea'
+export * from './Date'
+export * from './ItemTitle'
+export * from './StateIcons'
+export * from './TimelineItem'
