+++ src/views/pages/projectDetailsPage/components/timeline/index.ts
@@ -1,5 +1,5 @@
-export * from './components';
-export * from './helpers';
-export * from './DesignationItem';
-export * from './ImportItem';
-export * from './Timeline';
+export * from './components'
+export * from './helpers'
+export * from './DesignationItem'
+export * from './ImportItem'
+export * from './Timeline'
