--- src/views/pages/projectDetailsPage/index.ts
@@ -1,3 +1,0 @@
-export * from './components';
-export * from './sections';
-export * from './index';
